+++ tests/resources/TestClass.php
@@ -1,0 +1,6 @@
+<?php
+
+class Bar {
+    public function bar() {
+    }
+}
