+++ tests/resources/TestConstructor.php
@@ -1,0 +1,8 @@
+<?php
+
+class Bar {
+    public function bar() {
+    }
+    public function __construct() {
+    }
+}
