+++ tests/resources/TestNamespace.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Foo {
+    class Bar {
+        public function bar() {
+        }
+    }
+}
+namespace Bar {
+    class Bar {
+        public function bar() {
+        }
+    }
+}
